Implement double precision result buffer array on GPU (#89)
* Implement GPU double precision

* Use double for flux

* allow changing GPU result type in CMake

* Fix compile definition

* Bump viennacore

* Format

+++ gpu/include/raygTriangleGeometry.hpp
@@ -264,7 +264,7 @@ struct TriangleGeometry {
 
 #ifndef NDEBUG
     rayInternal::writeVTP(boundaryMesh, "triangleMesh_boundary.vtp",
-                          std::vector<float>());
+                          std::vector<double>());
 #endif
 
     return boundaryMesh;
